Move resources.ESBuild to js.Build
