+++ tpl/resources/resources.go
@@ -30,7 +30,6 @@ import (
 	"github.com/gohugoio/hugo/resources/resource_factories/bundler"
 	"github.com/gohugoio/hugo/resources/resource_factories/create"
 	"github.com/gohugoio/hugo/resources/resource_transformers/babel"
-	"github.com/gohugoio/hugo/resources/resource_transformers/esbuild"
 	"github.com/gohugoio/hugo/resources/resource_transformers/integrity"
 	"github.com/gohugoio/hugo/resources/resource_transformers/minifier"
 	"github.com/gohugoio/hugo/resources/resource_transformers/postcss"
@@ -65,7 +64,6 @@ func New(deps *deps.Deps) (*Namespace, error) {
 		postcssClient:   postcss.New(deps.ResourceSpec),
 		templatesClient: templates.New(deps.ResourceSpec, deps),
 		babelClient:     babel.New(deps.ResourceSpec),
-		esbuildClient:   esbuild.New(deps.BaseFs.Assets, deps.ResourceSpec),
 	}, nil
 }
 
@@ -80,7 +78,6 @@ type Namespace struct {
 	minifyClient    *minifier.Client
 	postcssClient   *postcss.Client
 	babelClient     *babel.Client
-	esbuildClient   *esbuild.Client
 	templatesClient *templates.Client
 }
 
@@ -302,25 +299,6 @@ func (ns *Namespace) Babel(args ...interface{}) (resource.Resource, error) {
 	}
 
 	return ns.babelClient.Process(r, options)
-
-}
-
-// ESBuild processes the given Resource with ESBuild.
-func (ns *Namespace) ESBuild(args ...interface{}) (resource.Resource, error) {
-	r, m, err := ns.resolveArgs(args)
-	if err != nil {
-		return nil, err
-	}
-	var options esbuild.Options
-	if m != nil {
-		options, err = esbuild.DecodeOptions(m)
-
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return ns.esbuildClient.Process(r, options)
 
 }
